correcao de notation no repository, metodos de repository foram trocados pelos de usecase no controller, correcao da arquitetura, troquei o tipo de id de int para Long e otimizei metodos de busca no repository

+++ src/main/java/com/cesurg/controlerio/core/domain/model/Cidade.java
@@ -1,19 +1,27 @@
 package com.cesurg.controlerio.core.domain.model;
 
 public class Cidade {
-    private int id;
+    private Long id;
     private String nome;
 
     public Cidade(){}
 
-    public Cidade(int id, String nome){
+    public Cidade(Long id, String nome){
         this.id = id;
         this.nome = nome;
     }
 
-    public int getId() {return id;}
-    public void setId(int id) {this.id = id;}
+    public Long getId() {
+        return id;
+    }
+    public void setId(Long id) {
+        this.id = id;
+    }
 
-    public String getNome() {return nome;}
-    public void setNome(String nome) {this.nome = nome;}
+    public String getNome() {
+        return nome;
+    }
+    public void setNome(String nome) {
+        this.nome = nome;
+    }
 }
